requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-import-export==3.2.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.6
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.3
tablib==3.4.0
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
